renamed web to components
