+++ modules/model_loader.py
@@ -3,6 +3,7 @@ from urllib.parse import urlparse
 from typing import Optional
 
 from bluemoon.utils.logly import logly
+from components.html import make_progress_html
 
 
 def load_file_from_url(
@@ -23,6 +24,10 @@ def load_file_from_url(
     cached_file = os.path.abspath(os.path.join(model_dir, file_name))
     if not os.path.exists(cached_file):
         logly.info(f'Downloading: "{file_name}" to {cached_file}\n')
-        from torch.hub import download_url_to_file
-        download_url_to_file(url, cached_file, progress=progress)
+        try:
+            from torch.hub import download_url_to_file
+            make_progress_html(0, f'Downloading: "{file_name}"')
+            download_url_to_file(url, cached_file, progress=progress)
+        except Exception as e:
+            logly.error(f'Failed to download "{file_name}" to {cached_file}: {e}')
     return cached_file
